moving to typed version wip
